Organization of menu bar component
