+++ src/app/_components/menu-bar/index.ts
@@ -1,0 +1,2 @@
+export * from './menu-bar.component';
+export * from './menubar-item/menubar-item.component';
