+++ src/app/_components/menu-bar/menu-bar.component.html
@@ -1,0 +1,6 @@
+<div class="ui horizontal pointing menu">
+  <menubar-item
+    [item]="item"
+    *ngFor="let item of items">
+  </menubar-item>
+</div>
