+++ src/app/_components/menu-bar/menu-bar.component.ts
@@ -1,0 +1,14 @@
+import {
+  Component,
+  Input
+} from '@angular/core';
+
+import { RouteDef } from '../../_models/index';
+
+@Component({
+  selector: 'menu-bar',
+  templateUrl: 'menu-bar.component.html'
+})
+export class MenuBarComponent {
+  @Input('items') items: RouteDef[];
+}
