+++ src/app/_components/menu-bar/menubar-item/menubar-item.component.html
@@ -1,0 +1,5 @@
+<a class="item"
+   [ngClass]="{ active: isActive() }"
+   [routerLink]="[item.path]">
+  {{ item.label }}
+</a>
