--- src/app/menu-bar/menu-bar.component.html
@@ -1,3 +1,0 @@
-<p>
-  menu-bar works!
-</p>
